Changed double to float. Added log bodies in console

+++ src/body.c
@@ -2,11 +2,11 @@
 #include <stdlib.h>
 #include "body.h"
 
-Vec2 *calc_force(Body *body1, Body *body2, double G){
-	double diffx_sqr = pow(body2->x - body1->x, 2);
-	double diffy_sqr = pow(body2->y - body1->y, 2);
-	double r_sqr = diffx_sqr + diffy_sqr;
-	double F = G * body1->m * body2->m / r_sqr;
+Vec2 *calc_force(Body *body1, Body *body2, float G){
+	float diffx_sqr = powf(body2->x - body1->x, 2.0);
+	float diffy_sqr = powf(body2->y - body1->y, 2.0);
+	float r_sqr = diffx_sqr + diffy_sqr;
+	float F = G * body1->m * body2->m / r_sqr;
 
 	Vec2 *F_vec = (Vec2 *)malloc(sizeof(Vec2));
 
@@ -14,14 +14,14 @@ Vec2 *calc_force(Body *body1, Body *body2, double G){
 		return NULL;
 	}
 
-	F_vec->x = F * sqrt(diffx_sqr / r_sqr);
-	F_vec->y = F * sqrt(diffy_sqr / r_sqr);
+	F_vec->x = F * sqrtf(diffx_sqr / r_sqr);
+	F_vec->y = F * sqrtf(diffy_sqr / r_sqr);
 
 	return F_vec;
 }
 
 // calc accel for body1
-Vec2 *calc_accel(Body *body1, Body *body2, double G){
+Vec2 *calc_accel(Body *body1, Body *body2, float G){
 	Vec2 *res = calc_force(body1, body2, G);
 
 	if (res == NULL){
@@ -42,7 +42,7 @@ Vec2 *calc_accel(Body *body1, Body *body2, double G){
 	return res;
 }
 
-void update_vel(Body *bodies, size_t len, double G){
+void update_vel(Body *bodies, size_t len, float G){
 	Vec2 *a = NULL;
 
 	for (size_t i = 0; i < len; i++){
@@ -65,7 +65,7 @@ void update_vel(Body *bodies, size_t len, double G){
 	}
 }
 
-void update_coords(Body *bodies, size_t len, double G){
+void update_coords(Body *bodies, size_t len, float G){
 	update_vel(bodies, len, G);
 
 	for (size_t i = 0; i < len; i++){
